Removes hack to prohibit stacking bufwrite events
This commit removes the hack I made to prohibit stacking `bufwrite`
events.

Previously, I set a global variable if the the user enables the plugin,
and checked this variable if the user tries to enable the plugin
consecutive times to prohibit stacking of events.

Now, I simply do it the right way, by utilizing an `augroup` which I
simply reset every time the user enables the plugin, before registering
any more events.